Add files via upload
I had the opportunity to participate in the Analytics Vidhya competition on demand forecasting for a meal delivery company that operates in multiple cities. The problem statement was to help the company's fulfilment centres with demand forecasting for the upcoming weeks, so they can plan their stock of raw materials and staffing needs accordingly.  
Most raw materials are replenished on a weekly basis, and with perishable items, accurate forecasting is essential for procurement planning. Moreover, staffing of the centres is also an area where accurate demand forecasts are helpful.  
The task was to predict the demand for the next 10 weeks (Weeks: 146-155) for the centre-meal combinations in the test set, given the following information:  
•	Historical data of demand for a product-centre combination (Weeks: 1 to 145).
•	Product (Meal) features such as category, sub-category, current price, and discount.
•	Information for fulfilment centre like centre area, city information etc. 
I approached the problem by exploring various techniques and models, and with continuous effort and perseverance, I was able to develop a solution that achieved high accuracy. My Python code file in PDF format showcases the steps I took to solve the problem and demonstrates the potential for future applications.